+++ marketplace/ui/src/components/MarketPlace/AddCart.js
@@ -418,7 +418,7 @@ const Checkout = ({ user }) => {
     if (isDone) {
       actions.addItemToCart(marketplaceDispatch, []);
       setTimeout(function () {
-        navigate(`/marketplace`);
+        navigate(`/`);
       }, 2000);
     }
   };