+++ marketplace/ui/src/components/Order/BoughtOrdersTable.js
@@ -226,10 +226,10 @@ const BoughtOrdersTable = ({ user, selectedDate, onDateChange, download, isAllOr
       // render: (text) => <p onClick={()=>{navigate(`${routes.MarketplaceUserProfile.url.replace(":commonName", text)}`, { state: { from: location.pathname } })}}>{text}</p>,
       render: (text) => (
         <a 
-          href={`${window.location.origin}/marketplace/profile/${encodeURIComponent(text)}`}
+          href={`${window.location.origin}/profile/${encodeURIComponent(text)}`}
           onClick={(e) => {
             e.preventDefault();
-            const userProfileUrl = `/marketplace/profile/${encodeURIComponent(text)}`;
+            const userProfileUrl = `/profile/${encodeURIComponent(text)}`;
       
             if (e.ctrlKey || e.metaKey) {
               // Open in a new tab if Ctrl/Cmd is pressed